+++ {{cookiecutter.directory_name}}/apps/api/decorators.py
@@ -9,12 +9,14 @@ def permission_required(perm):
     """
     Mark a view function to check specific user permission.
     """
+
     def decorator(func):
         def wrapper(request, *args, **kwargs):
             if request.user.has_perm(perm):
                 return func(request, *args, **kwargs)
             else:
-                raise ProblemDetailException(request, title=_('Permission denied.'), status_code=HTTPStatus.FORBIDDEN)
+                raise ProblemDetailException(request, title=_('Permission denied.'), status=HTTPStatus.FORBIDDEN)
 
         return wrapper
+
     return decorator