+++ {{cookiecutter.directory_name}}/apps/api/response.py
@@ -1,127 +1,141 @@
 import json
-from dataclasses import dataclass, field
+from dataclasses import dataclass
 from http import HTTPStatus
-from typing import Type, Union, List
+from typing import Optional, List, Type, TypeVar
 
 from django.conf import settings
-from django.db import models
-from django.db.models import QuerySet
 from django.core.paginator import Paginator, EmptyPage
-from django.http import HttpResponse
+from django.http import HttpResponse, HttpResponseRedirect
 from django.utils.translation import gettext as _
-from porcupine.base import Serializer
+from pydantic import RootModel, Field
 
 from apps.api.encoders import ApiJSONEncoder
-from apps.api.errors import ValidationException, ProblemDetailException
+from apps.api.errors import ProblemDetailException, DetailType, ValidationError, ProblemDetail
+from apps.core.serializers import Serializer
+
+ResponseType = TypeVar('ResponseType')
+
+
+class SingleResponseModel(Serializer):
+    response: ResponseType
+
+
+class PaginationModel(Serializer):
+    page: int
+    limit: Optional[int] = Field(default=None)
+    pages: int
+    total: int
+
+
+class PaginationResponseModel(Serializer):
+    items: RootModel[ResponseType]
+    metadata: PaginationModel
 
 
 @dataclass
 class Ordering:
-    columns: List[str] = field(default_factory=list)
+    columns: List[str]
 
     @classmethod
     def create_from_request(cls, request, aliases: dict = None) -> 'Ordering':
         columns = []
         aliases = aliases or {}
 
-        for column in request.GET.get('order_by', 'created_at').split(','):
-            column_name = column[1:] if column.startswith("-") else column
+        for column in request.GET.getlist('order_by', ['created_at']):
+            column_name = column[1:] if column.startswith('-') else column
             if column_name in aliases.keys():
-                alias_value = aliases[column_name]
-                if isinstance(alias_value, str):
-                    alias_value = [alias_value]
-                for val in alias_value:
-                    columns.append(
-                        f"-{val}" if column.startswith("-") else val
-                    )
+                columns.append(
+                    f'-{aliases[column_name]}' if column.startswith('-') else aliases[column_name]
+                )
             else:
                 columns.append(column)
-        return Ordering(columns)
+
+        result = Ordering(columns)
+        return result
 
     def __str__(self):
-        return ",".join(self.columns)
+        return ','.join(self.columns)
 
     def __repr__(self):
         return self.__str__()
 
 
 class GeneralResponse(HttpResponse):
-    def __init__(self, request, data: Union[models.Model, dict] = None, serializer: Type[Serializer] = None, **kwargs):
+    def __init__(self, request, data: Optional[ResponseType] = None, **kwargs):
         params = {}
         if data is not None:
             content_types = str(request.headers.get('accept', 'application/json'))
-            content_types = content_types.split(', ')
+            content_types = content_types.replace(' ', '').split(',')
             content_types = list(map(lambda r: r.split(';')[0], content_types))
 
             if any(x in ['*/*', 'application/json'] for x in content_types):
                 params['content_type'] = 'application/json'
-                params['content'] = json.dumps(data, cls=ApiJSONEncoder, serializer=serializer)
+                if isinstance(data, dict):
+                    params['content'] = json.dumps(data, cls=ApiJSONEncoder)
+                else:
+                    params['content'] = data.model_dump_json(by_alias=True)
             else:
                 params['content_type'] = 'application/json'
                 params['status'] = HTTPStatus.NOT_ACCEPTABLE
-                params['content'] = json.dumps({
-                    'message': _("Not Acceptable"),
-                    'metadata': {
-                        'available': [
-                            'application/json',
-                        ],
-                        'asked': ', '.join(content_types)
+                params['content'] = json.dumps(
+                    {
+                        'message': _('Not Acceptable'),
+                        'metadata': {
+                            'available': [
+                                'application/json',
+                            ],
+                            'asked': ', '.join(content_types)
+                        }
                     }
-                })
+                )
 
         kwargs.update(params)
         super().__init__(**kwargs)
 
 
 class SingleResponse(GeneralResponse):
-    def __init__(self, request, data=None, metadata: dict = None, **kwargs):
-        if data is None:
+    def __init__(self, request, data=None, serializer: Type[Serializer] = None, **kwargs):
+        if 'status' not in kwargs and data is None:
             kwargs['status'] = HTTPStatus.NO_CONTENT
-        else:
-            data = {
-                'response': data,
-            }
-            if metadata:
-                data['metadata'] = metadata
+        elif 'status' not in kwargs and data:
+            kwargs['status'] = HTTPStatus.OK
+
+        if data and serializer:
+            data = SingleResponseModel(
+                response=RootModel[serializer].model_validate(
+                    data,
+                    from_attributes=True,
+                    context={
+                        'request': request
+                    }
+                )
+            )
         super().__init__(request=request, data=data, **kwargs)
 
 
 class ErrorResponse(GeneralResponse):
-    def __init__(self, request, payload: dict, **kwargs):
+    def __init__(self, request, payload: ProblemDetail, **kwargs):
+        kwargs.setdefault('status', HTTPStatus.INTERNAL_SERVER_ERROR)
         super().__init__(request=request, data=payload, **kwargs)
-
-    @staticmethod
-    def create_from_exception(e: ProblemDetailException) -> 'ErrorResponse':
-        return ErrorResponse(e.request, e.payload, status=e.status, headers=e.extra_headers)
 
 
 class ValidationResponse(GeneralResponse):
-    def __init__(self, request, payload: dict, **kwargs):
-        data = {
-            'type': '/validation-error',
-            'title': 'Invalid request parameters',
-            'status': HTTPStatus.UNPROCESSABLE_ENTITY,
-            'errors': payload,
-        }
-
-        super().__init__(request, data, status=HTTPStatus.UNPROCESSABLE_ENTITY, **kwargs)
-
-    @staticmethod
-    def create_from_exception(e: ValidationException) -> 'ValidationResponse':
-        return ValidationResponse(e.request, e.payload, status=HTTPStatus.UNPROCESSABLE_ENTITY)
+    def __init__(self, request, payload: ValidationError, **kwargs):
+        kwargs.setdefault('content_type', 'application/problem+json')
+        kwargs.setdefault('status', HTTPStatus.UNPROCESSABLE_ENTITY)
+        super().__init__(request, payload, **kwargs)
 
 
 class PaginationResponse(GeneralResponse):
-    def __init__(self, request, qs, ordering: Ordering = None, extras: dict = None, **kwargs):
+    def __init__(self, request, qs, serializer: Type[Serializer], ordering: Ordering = None, **kwargs):
         kwargs.setdefault('content_type', 'application/json')
 
         # Ordering
-        if isinstance(qs, QuerySet):
-            ordering = ordering if ordering else Ordering.create_from_request(request)
-            qs = qs.order_by(str(ordering))
+        ordering = ordering if ordering else Ordering.create_from_request(request)
+        qs = qs.order_by(*ordering.columns)
 
+        # Pagination
         paginate = request.GET.get('paginate', 'true') == 'true'
-
         if paginate:
             limit = int(request.GET.get('limit', settings.PAGINATION['DEFAULT_LIMIT']))
             page = int(request.GET.get('page', 1))
@@ -136,7 +150,7 @@ class PaginationResponse(GeneralResponse):
                     title=_('Page not found'),
                     status=HTTPStatus.NOT_FOUND,
                     previous=e,
-                    detail_type='out_of_range',
+                    detail_type=DetailType.OUT_OF_RANGE,
                     detail=_('That page contains no results')
                 )
 
@@ -150,20 +164,27 @@ class PaginationResponse(GeneralResponse):
             num_pages = 1
             total = qs.count()
 
-        data = {}
-        if extras:
-            for key, item in extras.items():
-                data[key] = item
+        super().__init__(
+            request, PaginationResponseModel(
+                items=RootModel[List[serializer]].model_validate(
+                    list(items),
+                    from_attributes=True,
+                    context={
+                        'request': request
+                    }
+                ),
+                metadata=PaginationModel(
+                    page=page,
+                    limit=limit,
+                    pages=num_pages,
+                    total=total
+                )
+            ), **kwargs
+        )
 
-        data['items'] = items
-        data['metadata'] = {
-            'page': page,
-            'limit': limit,
-            'pages': num_pages,
-            'total': total
-        }
 
-        super().__init__(request, data, **kwargs)
+class SeeOtherResponse(HttpResponseRedirect):
+    status_code = 303
 
 
 __all__ = [
@@ -171,5 +192,6 @@ __all__ = [
     'ErrorResponse',
     'PaginationResponse',
     'ValidationResponse',
+    'SeeOtherResponse',
     'Ordering'
 ]