+++ {{cookiecutter.directory_name}}/apps/core/auth.py
@@ -1,40 +1,33 @@
 from http import HTTPStatus
-from typing import Optional
 
 from django.contrib.auth import get_user_model
 from django.contrib.auth.backends import ModelBackend
 from django.core.exceptions import ValidationError
-from django.conf import settings
 from django.utils import timezone
-from django.utils.text import slugify
 from django.utils.translation import gettext as _
 
-from apps.core.models import Token
-from apps.api.errors import ProblemDetailException
+from apps.core.models.token import Token
+from apps.api.errors import ProblemDetailException, UnauthorizedException
 
 User = get_user_model()
 
 
 class BearerBackend(ModelBackend):
-    @staticmethod
-    def _by_token(token) -> Optional[User]:
+    def authenticate(self, request, **kwargs) -> User:
         try:
-            token_obj = Token.objects.get(pk=token)
+            token = Token.objects.get(pk=kwargs['bearer'], expires_at__gte=timezone.now())
         except (Token.DoesNotExist, ValidationError):
-            return None
+            raise UnauthorizedException(request, detail=_('Invalid Bearer Token'))
 
-        token_obj.user.last_login = timezone.now()
-        token_obj.user.save()
+        if not self.user_can_authenticate(token.user):
+            raise ProblemDetailException(request, _('Inactive user.'), status=HTTPStatus.FORBIDDEN)
 
-        return token_obj.user
+        token.user.last_login = timezone.now()
+        token.user.save()
 
-    def authenticate(self, request, **kwargs) -> Optional[User]:
-        if kwargs.get('bearer'):
-            user = self._by_token(kwargs.get('bearer'))
-        else:
-            user = super().authenticate(request, **kwargs)
+        request.token = token
 
-        return user
+        return token.user
 
 
 class BasicBackend(ModelBackend):
@@ -42,12 +35,5 @@ class BasicBackend(ModelBackend):
         user = super().authenticate(request, **kwargs)
 
         if not user:
-            raise ProblemDetailException(
-                request,
-                _('Invalid credentials'),
-                status=HTTPStatus.UNAUTHORIZED,
-                extra_headers=(
-                    ('WWW-Authenticate', f'Bearer realm="{slugify(settings.INSTANCE_NAME)}"'),
-                )
-            )
+            raise UnauthorizedException(request)
         return user